Use localstorage to configure firebase.
Use a unique localstorage key per URL path to the single file version
of tiddlywiki to save and load the firestore configuration so that
serving the same wiki from different URLs permits different backend
storage.

Reconfigure serving so that /tw/* maps to individual users' wikis,
and /* maps to index.html to be replaced by a web app that handles
authentication, wiki creation, and sharing of wikis. Eventaually
users who arrive at the top level will be able to log in, find
their wikis, share their wikis, and create new ones.

+++ src/firebase.ts
@@ -18,25 +18,32 @@ import {
 // TODO: Add SDKs for Firebase products that you want to use
 // https://firebase.google.com/docs/web/setup#available-libraries
 
+const firebaseConfigTiddler = "$:/FirebaseConfig";
+const location =
+	typeof window !== "undefined" ? window.location.href : "build-time";
+const firebaseConfigStorage = `${firebaseConfigTiddler}|${location}}`;
 // Your web app's Firebase configuration
-const firebaseConfig = {
-	apiKey: "AIzaSyAAU8G6_I93RuQsfFdOf5wwdU4Wpn3cTXk",
-	authDomain: "tiddlywiki-a94cd.firebaseapp.com",
-	projectId: "tiddlywiki-a94cd",
-	storageBucket: "tiddlywiki-a94cd.appspot.com",
-	messagingSenderId: "251419323197",
-	appId: "1:251419323197:web:0e0ee30112d98099857354",
-};
+let firebaseConfig = undefined;
 
-// Initialize Firebase
-const app = initializeApp(firebaseConfig);
-const firestore = initializeFirestore(app, {
-	experimentalForceLongPolling: false, // isSafari(),
-});
+if (typeof window !== "undefined") {
+	const configString = window.localStorage.getItem(firebaseConfigStorage);
+	console.log(`FIREBASE: ${configString}`);
+	firebaseConfig = JSON.parse(configString);
+}
+
+const app = firebaseConfig ? initializeApp(firebaseConfig) : undefined;
+const firestore = app
+	? initializeFirestore(app, {
+			experimentalForceLongPolling: false, // isSafari(),
+	  })
+	: undefined;
 
 let initialLoadDone = false;
 export function initialLoadComplete() {
 	return initialLoadDone;
+}
+export function isConfiguration(id: string) {
+	return id === firebaseConfigTiddler;
 }
 
 function makeId(id: string) {
@@ -48,44 +55,63 @@ export function registerSyncCallback(
 		arg1: { modifications: string[]; deletions: string[] },
 	) => void,
 ) {
-	const tiddlers = collection(firestore, "tiddlers");
-	const unsub = onSnapshot(query(tiddlers), (querySnapshot) => {
-		const changed = querySnapshot.docChanges();
-		const modifications: string[] = [];
-		const deletions: string[] = [];
-		changed.forEach((docchange) => {
-			const title = docchange.doc.data().title;
-			if (docchange.type === "removed") {
-				deletions.push(title);
-			} else {
-				modifications.push(title);
-			}
-		});
-		if (modifications.length > 0 || deletions.length > 0) {
-			console.log(`Tiddler update mods=${JSON.stringify(modifications)} dels=${JSON.stringify(deletions)}`)
-			callback(null, { modifications, deletions });
-		}
-		initialLoadDone = true;
-	});
+	if (firestore) {
+		const tiddlers = collection(firestore, "tiddlers");
+		const unsub = onSnapshot(
+			query(tiddlers),
+			(querySnapshot) => {
+				const changed = querySnapshot.docChanges();
+				const modifications: string[] = [];
+				const deletions: string[] = [];
+				changed.forEach((docchange) => {
+					const title = docchange.doc.data().title;
+					if (docchange.type === "removed") {
+						deletions.push(title);
+					} else {
+						modifications.push(title);
+					}
+				});
+				if (modifications.length > 0 || deletions.length > 0) {
+					console.log(
+						`Tiddler update mods=${JSON.stringify(
+							modifications,
+						)} dels=${JSON.stringify(deletions)}`,
+					);
+					callback(null, { modifications, deletions });
+				}
+				initialLoadDone = true;
+			},
+			(err) => {
+				console.error(err);
+				alert(err);
+			},
+		);
+	}
 }
 export function storeTiddler(id: string, tiddlerData: any) {
-	if (id) {
+	if (id === firebaseConfigTiddler) {
+		window.localStorage.setItem(firebaseConfigStorage, tiddlerData.text);
+		window.location.reload();
+	}
+	if (id && firestore) {
 		const store = doc(firestore, `tiddlers/${makeId(id)}`);
 		return setDoc(store, tiddlerData);
 	}
 }
 
-export async function loadTiddler(id: string, callback: (arg0: null, arg1: DocumentData) => void) {
-	if (id) {
+export async function loadTiddler(
+	id: string,
+	callback: (arg0: null, arg1: DocumentData) => void,
+) {
+	if (id && firestore) {
 		const store = doc(firestore, `tiddlers/${makeId(id)}`);
 		const tiddler = await getDoc(store);
 		callback(null, tiddler.data());
 	}
 }
 
-
 export function deleteTiddler(id: string) {
-	if (id) {
+	if (id && firestore) {
 		const store = doc(firestore, `tiddlers/${makeId(id)}`);
 		return deleteDoc(store);
 	}
